refactor types placement
