+++ src/app/components/threads/ThreadsSidebar.tsx
@@ -2,7 +2,7 @@
 
 import ThreadRow from "./ThreadRow"
 import Toggle from "./Toggle"
-import { ThreadType } from "./types"
+import { ThreadType } from "../../types/threads"
 
 
 interface ThreadsSidebarProps {
